requirements.txt
opencv-python~=4.7.0.72
numpy~=1.24.3
matplotlib~=3.7.1
scikit-image==0.20.0
pandas~=2.0.1
Pillow~=9.5.0
Kivy~=2.1.0